refactor: split file into BlogPost, JobOffer and Vote
