+++ Imagine/Model/Imagine/BlogPost.swift
@@ -1,0 +1,22 @@
+//
+//  BlogPost.swift
+//  Imagine
+//
+//  Created by Malte Schoppe on 28.02.21.
+//  Copyright © 2021 Malte Schoppe. All rights reserved.
+//
+
+import Foundation
+
+class BlogPost {
+    var isCurrentProjectsCell = false
+    var title = ""
+    var subtitle = ""
+    var description = ""
+    var stringDate = ""
+    var category = ""
+    var poster = ""
+    var profileImageURL = ""
+    var imageURL = ""
+    var createDate = Date()
+}
