+++ Imagine/Model/Imagine/JobOffer.swift
@@ -1,0 +1,20 @@
+//
+//  JobOffer.swift
+//  Imagine
+//
+//  Created by Malte Schoppe on 28.02.21.
+//  Copyright © 2021 Malte Schoppe. All rights reserved.
+//
+
+import Foundation
+
+class JobOffer {
+    var title = ""
+    var cellText = ""
+    var descriptionText = ""
+    var documentID = ""
+    var stringDate = ""
+    var interested = 0
+    var category = ""
+    var createDate = Date()
+}
